rename make-test-settings
